backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
openai-whisper
groq
pydantic==2.5.0
pydantic-settings==2.1.0
playwright==1.40.0
supabase==2.0.3
